refactor: single errors file
